Random optimization

+++ Exiled.API/Features/Spawn/DynamicSpawnPoint.cs
@@ -28,7 +28,7 @@ namespace Exiled.API.Features.Spawn
         public SpawnLocationType Location { get; set; }
 
         /// <inheritdoc/>
-        public override float Chance { get; set; }
+        public override int Chance { get; set; }
 
         /// <inheritdoc/>
         [YamlIgnore]
